feature: 이달의 전시회 첫 번째 섹션 컴포넌트 분리
